+++ spec/Pim/Bundle/BaseConnectorBundle/Processor/Denormalization/Flat/GroupProcessorSpec.php
@@ -1,9 +1,10 @@
 <?php
 
-namespace spec\Pim\Bundle\BaseConnectorBundle\Processor\ArrayToObject\Flat;
+namespace spec\Pim\Bundle\BaseConnectorBundle\Processor\Denormalization;
 
 use Akeneo\Bundle\BatchBundle\Entity\StepExecution;
 use Akeneo\Bundle\BatchBundle\Item\InvalidItemException;
+use Akeneo\Bundle\StorageUtilsBundle\Doctrine\ObjectDetacherInterface;
 use PhpSpec\ObjectBehavior;
 use Pim\Bundle\CatalogBundle\Entity\Group;
 use Pim\Bundle\CatalogBundle\Entity\GroupType;
@@ -25,6 +26,7 @@ class GroupProcessorSpec extends ObjectBehavior
         GroupRepository $groupRepository,
         DenormalizerInterface $denormalizer,
         ValidatorInterface $validator,
+        ObjectDetacherInterface $detacher,
         StepExecution $stepExecution
     ) {
         $groupClass = 'Pim\Bundle\CatalogBundle\Entity\Group';
@@ -32,7 +34,9 @@ class GroupProcessorSpec extends ObjectBehavior
             $groupRepository,
             $denormalizer,
             $validator,
-            $groupClass
+            $detacher,
+            $groupClass,
+            'csv'
         );
         $this->setStepExecution($stepExecution);
         $validator->validate(Argument::any())->willReturn(new ConstraintViolationList());
@@ -155,6 +159,7 @@ class GroupProcessorSpec extends ObjectBehavior
         $groupRepository,
         $denormalizer,
         $validator,
+        $detacher,
         $stepExecution,
         Group $group,
         GroupType $type
@@ -184,6 +189,7 @@ class GroupProcessorSpec extends ObjectBehavior
             ->shouldBeCalled()
             ->willReturn($violations);
         $stepExecution->incrementSummaryInfo('skip')->shouldBeCalled();
+        $detacher->detach($group)->shouldBeCalled();
 
         $this
             ->shouldThrow(