+++ src/Pim/Bundle/BaseConnectorBundle/Processor/DummyProcessor.php
@@ -6,7 +6,7 @@ use Akeneo\Bundle\BatchBundle\Item\AbstractConfigurableStepElement;
 use Akeneo\Bundle\BatchBundle\Item\ItemProcessorInterface;
 
 /**
- * Dummy step, can be use to do nothing until you'll have concret implementation
+ * Dummy step, can be use to do nothing until you'll have concrete implementation
  *
  * @author    Nicolas Dupont <nicolas@akeneo.com>
  * @copyright 2013 Akeneo SAS (http://www.akeneo.com)