+++ src/Pim/Bundle/CatalogBundle/Updater/ProductTemplateUpdater.php
@@ -30,8 +30,6 @@ class ProductTemplateUpdater implements ProductTemplateUpdaterInterface
     public function update(ProductTemplateInterface $template, array $products)
     {
         $updates = $template->getValuesData();
-        // TODO unset identifier and axis updates and picture (not supported for now)
-
         foreach ($updates as $attributeCode => $values) {
             foreach ($values as $value) {
                 $this->productUpdater->setValue(